+++ src/main/java/br/edu/ufape/clickconsultas/negocios/servico/ServicoEnderecoMedico.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import br.edu.ufape.clickconsultas.dados.InterfaceColecaoEnderecoMedico;
 import br.edu.ufape.clickconsultas.negocios.modelo.EnderecoMedico;
+import br.edu.ufape.clickconsultas.negocios.servico.exception.ObjetoNaoEncontradoException;
 
 
 @Service
@@ -16,17 +17,28 @@ public class ServicoEnderecoMedico implements InterfaceServicoEnderecoMedico {
 	
 	public List<EnderecoMedico> buscarTodos() {
 		return colecaoEnderecoMedico.findAll();
+		
 	}
 		
-	public EnderecoMedico buscarPorId(long id) {
-		return colecaoEnderecoMedico.findById(id).orElse(null);
+	public EnderecoMedico buscarPorId(long id) throws ObjetoNaoEncontradoException {
+		EnderecoMedico endereco = colecaoEnderecoMedico.findById(id).orElse(null);
+		
+		if(endereco == null) 
+			throw new ObjetoNaoEncontradoException("o", "endereco");
+		
+		return endereco;
+			
 	}
 
 	public EnderecoMedico salvar(EnderecoMedico enderecoMedico) {
-		return colecaoEnderecoMedico.save(enderecoMedico);		
+		return colecaoEnderecoMedico.save(enderecoMedico);	
+		
 	}
 
-	public void remover(long id) {
-		colecaoEnderecoMedico.deleteById(id);
+	public void remover(long id) throws ObjetoNaoEncontradoException {
+		EnderecoMedico endereco = buscarPorId(id);
+		if(endereco != null)
+			colecaoEnderecoMedico.deleteById(id);
+		
 	}
 }
