+++ src/main/java/br/edu/ufape/clickconsultas/negocios/servico/ServicoHorarioAgendado.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import br.edu.ufape.clickconsultas.dados.InterfaceColecaoHorarioAgendado;
 import br.edu.ufape.clickconsultas.negocios.modelo.HorarioAgendado;
+import br.edu.ufape.clickconsultas.negocios.servico.exception.ObjetoNaoEncontradoException;
 
 @Service
 public class ServicoHorarioAgendado implements InterfaceServicoHorarioAgendado {
@@ -16,21 +17,38 @@ public class ServicoHorarioAgendado implements InterfaceServicoHorarioAgendado {
 	
 	public List<HorarioAgendado> buscarTodos() {
 		return colecaoHorarioAgendado.findAll();
+		
 	}
 	
-	public List<HorarioAgendado> BuscarPorData(LocalDate data) {
-		return colecaoHorarioAgendado.findByData(data);
+	public List<HorarioAgendado> BuscarPorData(LocalDate data) throws ObjetoNaoEncontradoException {
+		List<HorarioAgendado> lista = colecaoHorarioAgendado.findByData(data);
+		
+		if(lista.isEmpty())
+			throw new ObjetoNaoEncontradoException("o", "horario");
+		
+		return lista;
+		
 	}
 	
-	public HorarioAgendado buscarPorId(long id) {
-		return colecaoHorarioAgendado.findById(id).orElse(null);
+	public HorarioAgendado buscarPorId(long id) throws ObjetoNaoEncontradoException {
+		HorarioAgendado h = colecaoHorarioAgendado.findById(id).orElse(null);
+		
+		if(h == null)
+			throw new ObjetoNaoEncontradoException("o", "horario");
+		
+		return h;
+		
 	}
 
 	public HorarioAgendado salvar(HorarioAgendado horarioAgendado) {
 		return colecaoHorarioAgendado.save(horarioAgendado);		
 	}
 	
-	public void remover(long id) {
-		colecaoHorarioAgendado.deleteById(id);
+	public void remover(long id) throws ObjetoNaoEncontradoException {
+		HorarioAgendado h = buscarPorId(id);
+		
+		if(h != null) 
+			colecaoHorarioAgendado.deleteById(id);
+		
 	}
 }
